html/browsers/history/the-location-interface/context-for-window-open.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<!-- Incumbent Realm -->
<script>
function go() {
    frames[0].frames[0].open(
        "target-for-location.html", "target");
}
// https://html.spec.whatwg.org/C/#window-open-steps
async_test(t => {
  window.addEventListener("message", t.step_func_done(function(e) {
    assert_equals(
        e.data.location,
        new URL('resources/target-for-location.html', location.href).href,
        'Base URL should use the entry settings object');
    assert_equals(
        e.data.referrer,
        new URL('resources/middle-frame-for-location.html', location.href).href,
       'Referrer should use the entry settings object');
  }));
}, 'window.open()');
</script>
<iframe src="resources/middle-frame-for-location.html"></iframe>
